fix: equals and hashcode of several classes
