+++ src/test/java/dev/openfeature/sdk/ImmutableStructureTest.java
@@ -185,48 +185,4 @@ class ImmutableStructureTest {
 
         assertEquals(structure1, structure2);
     }
-
-    @Test
-    void unequalMutableStructuresAreNotEqual() {
-        MutableStructure m1 = new MutableStructure();
-        m1.add("key1", "val1");
-        MutableStructure m2 = new MutableStructure();
-        m2.add("key2", "val2");
-        assertNotEquals(m1, m2);
-    }
-
-    @Test
-    void equalMutableStructuresAreEqual() {
-        MutableStructure m1 = new MutableStructure();
-        m1.add("key1", "val1");
-        MutableStructure m2 = new MutableStructure();
-        m2.add("key1", "val1");
-        assertEquals(m1, m2);
-    }
-
-    @Test
-    void equalAbstractStructuresOfDifferentTypesAreEqual() {
-        MutableStructure m1 = new MutableStructure();
-        m1.add("key1", "val1");
-        HashMap<String, Value> map = new HashMap<>();
-        map.put("key1", new Value("val1"));
-        AbstractStructure m2 = new AbstractStructure(map) {
-            @Override
-            public Set<String> keySet() {
-                return attributes.keySet();
-            }
-
-            @Override
-            public Value getValue(String key) {
-                return attributes.get(key);
-            }
-
-            @Override
-            public Map<String, Value> asMap() {
-                return attributes;
-            }
-        };
-
-        assertEquals(m1, m2);
-    }
 }
